修复Leave实体类表名映射错误并优化代码规范
主要修复内容:
1. 修复Leave实体类表名映射错误
   - 将@TableName("leave")改为@TableName("emp_leave")
   - 解决程序运行时"Table 'leave' doesn't exist"错误

2. 优化SalaryPayment实体类方法命名
   - 将getSocialInsurancePersonal()改为getSocialInsurance()
   - 将setSocialInsurancePersonal()改为setSocialInsurance()
   - 使方法名与字段名保持一致，提高代码规范性

3. 更新数据库初始化脚本
   - 优化MySQL和Hive数据库初始化SQL
   - 完善数据插入脚本

4. 更新.gitignore文件
   - 添加node_modules排除规则
   - 添加测试文件和报告文件排除规则
   - 添加构建产物排除规则

测试结果:
- 编译测试: ✅ BUILD SUCCESS
- 打包测试: ✅ BUILD SUCCESS
- 启动测试: ✅ 启动成功 (5.43秒)
- 数据库对应: ✅ 14个表全部对应
- Leave功能: ✅ 修复完成

+++ backend/src/main/java/com/hr/datacenter/service/PerformanceEvaluationService.java
@@ -9,6 +9,8 @@ import com.hr.datacenter.mapper.mysql.PerformanceEvaluationMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 /**
  * 绩效评估Service
  *
@@ -57,11 +59,11 @@ public class PerformanceEvaluationService extends ServiceImpl<PerformanceEvaluat
         
         // 计算综合评分(自评40% + 上评60%)
         if (evaluation.getSelfScore() != null && evaluation.getSupervisorScore() != null) {
-            double finalScore = evaluation.getSelfScore() * 0.4 + evaluation.getSupervisorScore() * 0.6;
-            evaluation.setFinalScore(Math.round(finalScore * 100) / 100.0);
-            
+            double finalScore = evaluation.getSelfScore().doubleValue() * 0.4 + evaluation.getSupervisorScore().doubleValue() * 0.6;
+            evaluation.setFinalScore(BigDecimal.valueOf(Math.round(finalScore * 100) / 100.0));
+
             // 确定绩效等级
-            String level = determinePerformanceLevel(evaluation.getFinalScore());
+            String level = determinePerformanceLevel(evaluation.getFinalScore().doubleValue());
             evaluation.setPerformanceLevel(level);
         }
